+++ server/src/tcpClient.ts
@@ -44,6 +44,7 @@ export class TCPClient {
         log("tcpClient.connectToServer()");
         clearTimeout(this.reconnectTimer)
         this.socket = new net.Socket();
+        this.socket.setKeepAlive(true, 10000)
         this.socket.setTimeout(6000)
         this.socket.connect(this.port, this.host, () => {
             this.socket!.setTimeout(6000)