+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # 📜 Changelog
 
 
+#### 2025-03-06
+* Y turnout added.
+* Programming window implemented, currently works only with the DCC-EX command station.
+